Feature/heatbar

+++ Assets/src/Abilities/Utils/IAbility.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public interface IAbility {
 	
-	int Cost {get; set;}
+	float Cost {get; set;}
 	bool Executing {get; set;}
 	void Begin(ShipAction ship);
 	void Setup();
